+++ pynopegl-utils/pynopegl_utils/misc.py
@@ -19,84 +19,22 @@
 # under the License.
 #
 
-import inspect
 import json
 import os
 import os.path as op
 import pkgutil
-import platform
-import random
 import subprocess
 import tempfile
-from collections import namedtuple
-from dataclasses import asdict, dataclass, field
+from dataclasses import dataclass
 from fractions import Fraction
-from functools import wraps
-from typing import Any, Callable, Dict, List, Optional, Tuple
-
-from packaging.specifiers import SpecifierSet
-from packaging.version import Version
+from typing import Tuple
 
 import pynopegl as ngl
 
-
-def scene(controls: Optional[Dict[str, Any]] = None, compat_specs: Optional[str] = None):
-    def real_decorator(scene_func: Callable[..., ngl.Node]) -> Callable[..., SceneInfo]:
-        @wraps(scene_func)
-        def func_wrapper(scene_cfg: Optional[SceneCfg] = None, **extra_args):
-            version = Version(ngl.__version__)
-            if compat_specs and version not in SpecifierSet(compat_specs):
-                raise Exception(
-                    f"{scene_func.__name__} needs libnopegl{compat_specs} but libnopegl is currently at version {version}"
-                )
-
-            if scene_cfg is None:
-                scene_cfg = SceneCfg()
-            root = scene_func(scene_cfg, **extra_args)
-            scene = ngl.Scene.from_params(root, scene_cfg.duration, scene_cfg.framerate, scene_cfg.aspect_ratio)
-            return SceneInfo(
-                scene=scene,
-                backend=scene_cfg.backend,
-                samples=scene_cfg.samples,
-                clear_color=scene_cfg.clear_color,
-                files=scene_cfg.files,
-            )
-
-        # Construct widgets specs
-        widgets_specs = []
-        func_specs = inspect.getfullargspec(scene_func)
-        if controls is not None and func_specs.defaults:
-            nb_optionnals = len(func_specs.defaults)
-            for i, key in enumerate(func_specs.args[-nb_optionnals:]):
-                # Set controller defaults according to the function prototype
-                control = controls.get(key)
-                if control is not None:
-                    default = func_specs.defaults[i]
-                    ctl_id = control.__class__.__name__
-                    ctl_data = control._asdict()
-                    widgets_specs.append((key, default, ctl_id, ctl_data))
-
-        # Transfers the widget specs to the UI.
-        # We could use the return value but it's better if the user can still
-        # call its decorated scene function transparently inside his own code
-        # without getting garbage along the return value.
-        func_wrapper.widgets_specs = widgets_specs
-
-        # Flag the scene as a scene function so it's registered in the UI.
-        func_wrapper.iam_a_ngl_scene_func = True
-
-        return func_wrapper
-
-    return real_decorator
-
-
-scene.Range = namedtuple("Range", "range unit_base", defaults=([0, 1], 1))
-scene.Vector = namedtuple("Vector", "n minv maxv", defaults=(None, None))
-scene.Color = namedtuple("Color", "")
-scene.Bool = namedtuple("Bool", "")
-scene.File = namedtuple("File", "filter", defaults=("",))
-scene.List = namedtuple("List", "choices")
-scene.Text = namedtuple("Text", "")
+# Temporary kept for compatibility
+scene = ngl.scene
+SceneCfg = ngl.SceneCfg
+SceneInfo = ngl.SceneInfo
 
 
 @dataclass(frozen=True)
@@ -137,15 +75,6 @@ class MediaInfo:
         )
 
 
-@dataclass
-class SceneInfo:
-    scene: ngl.Scene
-    backend: str
-    samples: int
-    clear_color: Tuple[float, float, float, float]
-    files: List[str]
-
-
 def get_nopegl_tempdir() -> str:
     tmpdir = op.join(tempfile.gettempdir(), "nopegl")
     os.makedirs(tmpdir, exist_ok=True)
@@ -160,29 +89,6 @@ MEDIA_FILES_DB = dict(
     rooster="Unsplash-Michael_Anfang-Rooster-cropped.jpg",
     panda="Unsplash-Romane_Gautun-Red_Panda-cropped.jpg",
 )
-
-
-@dataclass
-class SceneCfg:
-    aspect_ratio: Tuple[int, int] = (16, 9)
-    duration: float = 30.0
-    framerate: Tuple[int, int] = (60, 1)
-    backend: str = "opengl"
-    samples: int = 0
-    system: str = platform.system()
-    files: List[str] = field(default_factory=list)
-    clear_color: Tuple[float, float, float, float] = (0.0, 0.0, 0.0, 1.0)
-
-    def __post_init__(self):
-        # Predictible random number generator
-        self.rng = random.Random(0)
-
-    @property
-    def aspect_ratio_float(self) -> float:
-        return self.aspect_ratio[0] / self.aspect_ratio[1]
-
-    def as_dict(self) -> Dict[str, Any]:
-        return asdict(self)
 
 
 def get_shader(name: str) -> str: